requirements2.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cycler==0.11.0
dnspython==2.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
fonttools==4.37.3
gunicorn==19.7.1
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
py-bcrypt==0.4
pycparser==2.21
pymongo==4.2.0
pyparsing==3.0.9
python-bcrypt==0.3.2
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.8.1
